refactor: property 주입
