+++ nextstep-jdbc/src/main/java/nextstep/jdbc/JdbcTemplate.java
@@ -3,6 +3,7 @@ package nextstep.jdbc;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.sql.DataSource;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -12,12 +13,17 @@ import java.util.List;
 
 public class JdbcTemplate {
     private static final Logger log = LoggerFactory.getLogger(JdbcTemplate.class);
+    private final DataSource dataSource;
+
+    public JdbcTemplate(DataSource dataSource) {
+        this.dataSource = dataSource;
+    }
 
     private static void values(PreparedStatement pstmt) {
     }
 
     public int update(String query, PreparedStatementSetter setter) {
-        try (Connection con = ConnectionManager.getConnection();
+        try (Connection con = dataSource.getConnection();
              PreparedStatement pstmt = con.prepareStatement(query)) {
             setter.values(pstmt);
             return pstmt.executeUpdate();
@@ -32,7 +38,7 @@ public class JdbcTemplate {
     }
 
     public <T> List<T> query(String query, RowMapper<T> rowMapper, PreparedStatementSetter setter) {
-        try (Connection con = ConnectionManager.getConnection();
+        try (Connection con = dataSource.getConnection();
              PreparedStatement pstmt = con.prepareStatement(query)) {
             setter.values(pstmt);
             return getQueryResults(rowMapper, pstmt);
@@ -61,7 +67,7 @@ public class JdbcTemplate {
     }
 
     public <T> T queryForObject(String query, RowMapper<T> rowMapper, PreparedStatementSetter setter) {
-        try (Connection con = ConnectionManager.getConnection();
+        try (Connection con = dataSource.getConnection();
              PreparedStatement pstmt = con.prepareStatement(query)) {
             setter.values(pstmt);
             return getQueryResult(rowMapper, pstmt);
