+++ slipp/src/main/java/slipp/controller/UpdateFormUserController.java
@@ -4,12 +4,13 @@ import nextstep.jdbc.JdbcTemplate;
 import nextstep.mvc.asis.Controller;
 import slipp.dao.UserDao;
 import slipp.domain.User;
+import slipp.support.db.ConnectionManager;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 public class UpdateFormUserController implements Controller {
-    private final UserDao userDao = new UserDao(new JdbcTemplate());
+    private final UserDao userDao = new UserDao(new JdbcTemplate(ConnectionManager.getDataSource()));
 
     @Override
     public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
