+++ slipp/src/main/java/slipp/support/db/ConnectionManager.java
@@ -1,23 +1,35 @@
-package nextstep.jdbc;
+package slipp.support.db;
 
 import org.apache.commons.dbcp2.BasicDataSource;
 
 import javax.sql.DataSource;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.InputStream;
 import java.sql.Connection;
 import java.sql.SQLException;
+import java.util.Properties;
 
 public class ConnectionManager {
-    private static final String DB_DRIVER = "org.h2.Driver";
-    private static final String DB_URL = "jdbc:h2:mem:jwp-framework";
-    private static final String DB_USERNAME = "sa";
-    private static final String DB_PW = "";
+    private static final String DB_DRIVER_KEY = "jdbc.driverClass";
+    private static final String DB_URL_KEY = "jdbc.url";
+    private static final String DB_USERNAME_KEY = "jdbc.username";
+    private static final String DB_PW_KEY = "jdbc.password";
 
     public static DataSource getDataSource() {
         BasicDataSource ds = new BasicDataSource();
-        ds.setDriverClassName(DB_DRIVER);
-        ds.setUrl(DB_URL);
-        ds.setUsername(DB_USERNAME);
-        ds.setPassword(DB_PW);
+        try (InputStream input = new FileInputStream("src/main/resources/db.properties")) {
+            Properties prop = new Properties();
+
+            prop.load(input);
+            ds.setDriverClassName(prop.getProperty(DB_DRIVER_KEY));
+            ds.setUrl(prop.getProperty(DB_URL_KEY));
+            ds.setUsername(prop.getProperty(DB_USERNAME_KEY));
+            ds.setPassword(prop.getProperty(DB_PW_KEY));
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
         return ds;
     }
 
@@ -29,3 +41,4 @@ public class ConnectionManager {
         }
     }
 }
+
