+++ slipp/src/test/java/slipp/dao/UserDaoTest.java
@@ -1,6 +1,5 @@
 package slipp.dao;
 
-import nextstep.jdbc.ConnectionManager;
 import nextstep.jdbc.JdbcTemplate;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -9,6 +8,7 @@ import org.springframework.jdbc.datasource.init.DatabasePopulatorUtils;
 import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator;
 import slipp.domain.User;
 import slipp.dto.UserUpdatedDto;
+import slipp.support.db.ConnectionManager;
 
 import java.util.List;
 
@@ -19,7 +19,7 @@ public class UserDaoTest {
 
     @BeforeEach
     public void setup() {
-        userDao = new UserDao(new JdbcTemplate());
+        userDao = new UserDao(new JdbcTemplate(ConnectionManager.getDataSource()));
         ResourceDatabasePopulator populator = new ResourceDatabasePopulator();
         populator.addScript(new ClassPathResource("jwp.sql"));
         DatabasePopulatorUtils.execute(populator, ConnectionManager.getDataSource());
